TV - barchart refactoring
